+++ lms/djangoapps/instructor/views/api_urls.py
@@ -3,7 +3,7 @@
 Instructor API endpoint urls.
 """
 
-from django.urls import include, path, re_path
+from django.urls import path, re_path
 
 from lms.djangoapps.instructor.views import api, gradebook_api, api_v2
 from openedx.core.constants import COURSE_ID_PATTERN
@@ -21,9 +21,6 @@ v1_api_urls = [
 ]
 
 v2_api_urls = [
-    path(
-        '', include('lms.djangoapps.instructor.api.urls'),
-    ),
     re_path(
         rf'^courses/{COURSE_ID_PATTERN}$',
         api_v2.CourseMetadataView.as_view(),
@@ -43,7 +40,12 @@ v2_api_urls = [
         rf'^courses/{COURSE_ID_PATTERN}/graded_subsections$',
         api_v2.GradedSubsectionsView.as_view(),
         name='graded_subsections'
-    )
+    ),
+    re_path(
+        rf'^courses/{COURSE_ID_PATTERN}/ora$',
+        api_v2.ORAView.as_view(),
+        name='ora_assessments'
+    ),
 ]
 
 urlpatterns = [
